deleted unecessary files and refactored code
